+++ src/test_pages/BasePage.ts
@@ -1,4 +1,5 @@
 import { Locator, Page, expect, test } from '@playwright/test'
+import { setBrowserTime } from 'tests/utils'
 
 export abstract class BasePage {
   constructor(protected page: Page) {}
@@ -7,6 +8,10 @@ export abstract class BasePage {
     await test.step(`Validating that a correct value of URL is ${url}`, async () => {
       await expect(this.page).toHaveURL(url)
     })
+  }
+
+  async setFakeTime(date: Date) {
+    await setBrowserTime(date, this.page)
   }
 
   protected async clickOnElement(element: Locator, timeout?: number) {
@@ -51,27 +56,5 @@ export abstract class BasePage {
       options = await this.page.getByRole('option').all()
     })
     return options
-  }
-
-  async setFakeTime(date: Date) {
-    const fakeNow = date.valueOf()
-
-    // Update the Date accordingly
-    await this.page.addInitScript(`{
-      // Extend Date constructor to default to fakeNow
-      Date = class extends Date {
-        constructor(...args) {
-          if (args.length === 0) {
-            super(${fakeNow});
-          } else {
-            super(...args);
-          }
-        }
-      }
-      // Override Date.now() to start from fakeNow
-      const __DateNowOffset = ${fakeNow} - Date.now();
-      const __DateNow = Date.now;
-      Date.now = () => __DateNow() + __DateNowOffset;
-    }`)
   }
 }
