+++ tests/utils.ts
@@ -4,6 +4,7 @@ import * as path from 'path'
 import * as crypto from 'crypto'
 import moment from 'moment'
 import { Matcher, test as baseTest, customMatcher } from 'playwright-advanced-har'
+import { Page } from '@playwright/test'
 
 const istanbulCLIOutput = path.join(process.cwd(), '.nyc_output')
 
@@ -37,6 +38,28 @@ export const test = baseTest.extend({
 
 export function getPastDate(): Date {
   return moment('2024-02-12 15:00:00').toDate()
+}
+
+export async function setBrowserTime(date: Date, page: Page) {
+  const fakeNow = date.valueOf()
+
+  // Update the Date accordingly
+  await page.addInitScript(`{
+    // Extend Date constructor to default to fakeNow
+    Date = class extends Date {
+      constructor(...args) {
+        if (args.length === 0) {
+          super(${fakeNow});
+        } else {
+          super(...args);
+        }
+      }
+    }
+    // Override Date.now() to start from fakeNow
+    const __DateNowOffset = ${fakeNow} - Date.now();
+    const __DateNow = Date.now;
+    Date.now = () => __DateNow() + __DateNowOffset;
+  }`)
 }
 
 export function urlMatcher(): Matcher {
